add script parser

+++ src/main/kotlin/com/fujitsu/labs/virtualhome/Graph.kt
@@ -15,8 +15,8 @@ data class Node(
     val category: String? = null,
     val class_name: String? = null,
     val prefab_name: String? = null,
-    val obj_transform: Map<String, List<Double>>? = null,
-    val bounding_box: Map<String, List<Double>>? = null,
+    val obj_transform: Map<String, List<Float>>? = null,
+    val bounding_box: Map<String, List<Float>>? = null,
     val properties: List<String>? = null,
     val states: List<String>? = null
 )
